requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
pydantic==2.11.7
python-multipart==0.0.20
httpx==0.28.1
langchain-openai==0.3.24
langchain==0.3.25
langsmith==0.3.45
pydantic-settings==2.9.1
pytest==8.4.1
python-dotenv==1.1.0
